Heal pack has been added
It works like this: When a player's spaceship lands on a health pack, it heals them for 20HP (you can change the amount in the code) when their turn ends. In the command code, you will see messaging saying, "Checking if a spaceship has laned on a power up..." Also, enemieships can't get any powerups. That is why the checking code is placed so they can't get anything. If one lands on it, it won't heal them. And when a player spaceship lands on it, you will get a message saying, "Power up was detected and it went to Blue Mothership!" and it says this after, "Blue Mothership has been healed to 170 HP!"